% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_main_func.R
\name{compute_dendro_sp}
\alias{compute_dendro_sp}
\title{Compute dendrological variable for one species at the plot scale}
\usage{
compute_dendro_sp(NFI_tree, species_name)
}
\arguments{
\item{NFI_tree}{The data.table storing the tree data, usually \strong{NFI_tree} retured by {\link[=get_NFI]{get_NFI}}}

\item{species_name}{A Character indicating the NFI species code, look for "espar" in the metadata.csv for more informations}
}
\value{
a data.table of 2 dendrological variable (basal area of the species, and its number of stem per hectares) and the plot id idp
}
\description{
Compute dendrological variable for one species at the plot scale
}
\examples{
\dontrun{

NFI_douglas<-compute_dendro_sp(NFI_tree,"64")
head(NFI_douglas)
}
}
\references{
https://inventaire-forestier.ign.fr/
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
