% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_main_func.R
\name{download_NFI}
\alias{download_NFI}
\title{Download the latest NFI}
\usage{
download_NFI(dir = getwd(), nfi_url = get_latest_url())
}
\arguments{
\item{nfi_url}{The url}
}
\value{
The path to where the data have been downloaded
}
\description{
Download the latest NFI
}
\details{
The raw downloaded data are stored inside the NFI_data/Raw_data folder
}
\note{

}
\examples{
\dontrun{
download_NFI(dir=getwd())
}
}
\references{
https://inventaire-forestier.ign.fr/
}
\seealso{
{\link[=get_NFI]{get_NFI}}
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
