% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_main_func.R
\name{get_NFI}
\alias{get_NFI}
\title{Get and format the NFI}
\usage{
get_NFI(
  dir = getwd(),
  nfi_url = get_latest_url(),
  write_csv = F,
  export_to_env = T,
  visit = 1
)
}
\arguments{
\item{dir}{The path indicated where the raw and formated NFI is gonna be stored}

\item{nfi_url}{The NFI download NFI, a default value is provided}

\item{write_csv}{Logical, should the NFI be return in .csv files in the NFI_data folder}

\item{export_to_env}{Logical, should the NFI be directly imported into the global environement}

\item{visit}{1 to get the first NFI measure (default), 2 for the revisist or c(1,2) for both}
}
\value{
If \code{export_to_env}  is FALSE, a list of 6 data.table
}
\description{
Get and format the NFI
}
\details{
The raw downloaded data are stored inside the NFI_data/Raw_data folder
The NFI is seperated into different csv files or R data.table, that you can merge by idp
\itemize{
\item{\strong{NFI_plot_info}}{ The metadata such as position, or time of measurment of a NFI plot, plot are identified with a unique ID called idp}
\item{\strong{NFI_ecologie}}{ Informations regarding soil, understory cover observations}
\item{\strong{NFI_arbres}}{ Living tree measurements and informations}
\item{\strong{NFI_dendro}}{ Dendrologic variable computed with NFI_arbres}
\item{\strong{NFI_cover}}{ Relative and absolute cover of every tree species recorded}
\item{\strong{NFI_flora}}{ Florisitic surveys of the NFI plot}
}
}
\examples{
\dontrun{
get_NFI(dir=getwd(),write_csv=F,export_to_env=T,visit=1)
}
}
\references{
https://inventaire-forestier.ign.fr/
}
\author{
J Borderieux (jeremy.borderieux@agroparistech.fr)
}
